+++ subsys/bluetooth/controller/ll_sw/ull_iso_internal.h
@@ -1,0 +1,8 @@
+/*
+ * Copyright (c) 2020 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+int ull_iso_init(void);
+int ull_iso_reset(void);
